Extracing character into tabs
