+++ src/features/Character/components/Bags/index.js
@@ -8,7 +8,7 @@ import actions from 'features/Gw2/actions';
 function mapStateToProps (state) {
   return {
     items: state.items,
-    bags: state.characters.data[state.characters.selected].bags,
+    bags: (state.characters.data[state.characters.selected] || {}).bags,
   };
 }
